requirements.txt
faster-whisper>=1.1.0
opencc-python-reimplemented>=0.1.7
tqdm>=4.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
pymysql>=1.1.0
paramiko>=3.0.0
